Move all project-link related methods to a concern
Also implements .links_to_remote?
